implemment UIPopoverPresentationController.delegate mechanism.
previously the popover delegation methods were not being handled
because the related delegate class in iOS8
("UIPopoverPresentationControllerDelegate") does not exist in
iOS<=7. Furtunately this new delegate is very similar to
UIPopoverControllerDelegate and therefore it seemed possible
to catch calls to the old delegate type and pass them to
the new one.

JVPopoverPresentationController is the glue between old and
new delegate type. It facedes UIPopoverController to implement
properties of the new type UIPopoverPresentationController.