make this library MRC (manual reference counting) compliant.
There are still applications out there that use MRC because of
legacy code. In order to support these applications some code
has been regrouped and "release" messages added where necessary.
With a couple of definitions I tried to keep the impact as low
as possible and make it not too hard to maintain compatibility
with ARC.